+++ decidim-admin/app/models/decidim/admin/abilities/participatory_process_admin.rb
@@ -9,18 +9,9 @@ module Decidim
       # Admin engine.
       #
       # This ability will not apply to organization admins.
-      class ParticipatoryProcessAdmin
-        include CanCan::Ability
-
-        def initialize(user, context)
-          @user = user
-          @context = context
-
-          define_abilities if @user && !@user.admin? && has_manageable_processes?
-        end
-
+      class ParticipatoryProcessAdmin < Decidim::Abilities::ParticipatoryProcessAdmin
         def define_abilities
-          can :read, :admin_dashboard
+          super
 
           can :manage, ParticipatoryProcess do |process|
             can_manage_process?(process)
@@ -28,11 +19,11 @@ module Decidim
 
           cannot :create, ParticipatoryProcess
           cannot :destroy, ParticipatoryProcess
-
-          define_participatory_process_abilities if current_participatory_process && can_manage_process?(current_participatory_process)
         end
 
         def define_participatory_process_abilities
+          super
+
           can :manage, Feature do |feature|
             can_manage_process?(feature.participatory_process)
           end
@@ -56,22 +47,6 @@ module Decidim
           can :manage, ParticipatoryProcessStep do |step|
             can_manage_process?(step.participatory_process)
           end
-        end
-
-        def current_participatory_process
-          @current_participatory_process ||= @context[:current_participatory_process]
-        end
-
-        def participatory_processes_with_admin_role
-          @participatory_processes ||= Decidim::ParticipatoryProcessesWithUserRole.for(@user, :admin)
-        end
-
-        def can_manage_process?(process)
-          participatory_processes_with_admin_role.include? process
-        end
-
-        def has_manageable_processes?
-          participatory_processes_with_admin_role.any?
         end
       end
     end
