+++ decidim-admin/app/models/decidim/admin/abilities/participatory_process_collaborator.rb
@@ -5,24 +5,14 @@ module Decidim
     module Abilities
       # Defines the abilities for a collaborator user in the admin
       # section. Intended to be used with `cancancan`.
-      class ParticipatoryProcessCollaborator
-        include CanCan::Ability
+      class ParticipatoryProcessCollaborator < Decidim::Abilities::ParticipatoryProcessCollaborator
+        def define_abilities
+          super
 
-        def initialize(user, _context)
-        #   @user = user
-
-        #   return unless user && !user.admin?
-
-        #   can :manage, :admin_dashboard
-
-        #   can :preview, ParticipatoryProcess do |process|
-        #     participatory_processes.include?(process)
-        #   end
+          can [:read, :preview], ParticipatoryProcess do |process|
+            can_manage_process?(process)
+          end
         end
-
-        # def participatory_processes
-        #   @participatory_processes ||= Decidim::ParticipatoryProcessesWithUserRole.for(@user, :collaborator)
-        # end
       end
     end
   end
