+++ decidim-core/app/models/decidim/abilities/participatory_process_admin.rb
@@ -5,26 +5,10 @@ module Decidim
     # Defines the abilities for a participatory process admin. Intended to be
     # used with `cancancan`.
     # This ability will not apply to organization admins.
-    class ParticipatoryProcessAdmin
-      include CanCan::Ability
-
-      def initialize(user, context)
-        @user = user
-        @context = context
-
-        define_abilities if @user && !@user.admin? && has_manageable_processes?
-      end
-
-      def define_abilities
-        can :read, :admin_dashboard
-      end
-
-      def participatory_processes_with_admin_role
-        @participatory_processes ||= Decidim::ParticipatoryProcessesWithUserRole.for(@user, :admin)
-      end
-
-      def has_manageable_processes?
-        participatory_processes_with_admin_role.any?
+    class ParticipatoryProcessAdmin < ParticipatoryProcessUserRole
+      # Overrides ParticipatoryProcessUserRole role method
+      def role
+        :admin
       end
     end
   end
